% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cov_r2C}
\alias{Cov_r2C}
\title{Sum of covariances of squared correlations}
\usage{
Cov_r2C(n, R, E)
}
\arguments{
\item{n}{Degrees of freedom (not sample sizes); numeric of length 1 or more.}

\item{R}{Population correlation matrix; assumed to be validly constructed;
numeric of length \code{p * p}}

\item{E}{Matrix of expectation of correlation coefficients corresponding to
the upper triangular of \code{R}; numeric of length
\code{p * (p - 1) / 2 * length(n)}.}
}
\value{
A numeric vector of \eqn{\sum 2 Cov(r_{ij}, r_{kl})},
corresponding to \code{n}.
}
\description{
\code{Cov.r2C()} calculates (twice) sum of covariances of
all non-redundant pairs of squared correlation coefficients.
This is an internal C++ function to be used in within \code{AVar.VRr_pfc()}
with the package \code{Rcpp}.
}
\details{
Primarily for internal use.

It is assumed that \code{E} has the form of
\code{matrix(sapply(n, Exv.r1, x = R[upper.tri(R)]), ncol = length(n))}.

\code{R} and \code{E} are read as NumericVector rather than
NumericMatrix to save a modest amount of time
}
\seealso{
\link{Exv.VXx}, \link{AVar.VRr_xx}
}
