% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{digit}
\alias{digit}
\title{Obtain number of digits}
\usage{
digit(x)
}
\arguments{
\item{x}{The object whose number of digits is to be obtained.  Length can
be >1.  Integer or integer-like numeric are assumed.}
}
\value{
The number of digits, as an integer vector
}
\description{
Internal utility function to obtain the number of digit of an integer.  For
internal use in \code{AVar.VRr_pfx()} functions (and others).
}
