% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{divInd}
\alias{divInd}
\title{Divide index vector}
\usage{
divInd(b, Max = 2e+06)
}
\arguments{
\item{b}{Vector to be divided into a list}

\item{Max}{Maximum acceptable length of vector that results from the divided list}
}
\value{
A list
}
\description{
Utility function to divide a vector of index to a list.
To be used internally in \code{AVar.VRr_pfd()}.
}
\details{
The argument \code{Max} defines the maximum of \code{length(bd[[i]])} in
\code{AVar.VRr_pfd}, which roughly defines the amount of RAM required
(see description of that function).
}
