% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_rx.R
\name{Exv.rx}
\alias{Exv.rx}
\alias{Exv.r1}
\alias{Exv.r2}
\alias{Exv.r3}
\alias{Exv.r3d}
\alias{Exv.r4}
\alias{Exv.r4d}
\alias{Var.r1}
\alias{Var.r2}
\alias{Cm3.r1}
\alias{Cm4.r1}
\alias{AExv.r1}
\alias{AVar.r1}
\alias{ACm3.r1}
\alias{ACm4.r1}
\alias{AExv.r2}
\alias{AExv.r3}
\alias{AExv.r4}
\alias{AVar.r2}
\alias{ACov.r1}
\title{Moments of correlation coefficients}
\usage{
Exv.r1(n, x, tol = .Machine$double.eps * 100, tol.hg = 0, maxiter = 2000, ...)

Exv.r2(
  n,
  x,
  do.square = TRUE,
  tol = .Machine$double.eps * 100,
  tol.hg = 0,
  maxiter = 2000,
  ...
)

Exv.r3(n, x, tol = .Machine$double.eps * 100, tol.hg = 0, maxiter = 2000, ...)

Exv.r3d(n, x, tol = .Machine$double.eps * 100, tol.hg = 0, maxiter = 2000, ...)

Exv.r4(
  n,
  x,
  do.square = TRUE,
  tol = .Machine$double.eps * 100,
  tol.hg = 0,
  maxiter = 2000,
  ...
)

Exv.r4d(
  n,
  x,
  do.square = TRUE,
  tol = .Machine$double.eps * 100,
  tol.hg = 0,
  maxiter = 2000,
  ...
)

Var.r1(n, x, tol = .Machine$double.eps * 100, tol.hg = 0, maxiter = 2000, ...)

Var.r2(
  n,
  x,
  do.square = TRUE,
  tol = .Machine$double.eps * 100,
  tol.hg = 0,
  maxiter = 2000,
  ...
)

Cm3.r1(n, x, tol = .Machine$double.eps * 100, tol.hg = 0, maxiter = 2000, ...)

Cm4.r1(n, x, tol = .Machine$double.eps * 100, tol.hg = 0, maxiter = 2000, ...)

AExv.r1(n, x, order. = 2, return_terms = FALSE, ...)

AVar.r1(n, x, order. = 2, return_terms = FALSE, ...)

ACm3.r1(n, x, order. = 3, return_terms = FALSE, ...)

ACm4.r1(n, x, order. = 3, return_terms = FALSE, ...)

AExv.r2(n, x, order. = 2, return_terms = FALSE, ...)

AExv.r3(n, x, order. = 2, return_terms = FALSE, ...)

AExv.r4(n, x, order. = 2, return_terms = FALSE, ...)

AVar.r2(n, x, order. = 2, return_terms = FALSE, ...)

ACov.r1(
  n,
  R,
  ind1 = c(1, 2),
  ind2 = c(1, 2),
  i = ind1[1],
  j = ind1[2],
  k = ind2[1],
  l = ind2[2],
  rij = R[i, j],
  rik = R[i, k],
  ril = R[i, l],
  rjk = R[j, k],
  rjl = R[j, l],
  rkl = R[k, l],
  ...
)
}
\arguments{
\item{n}{Degrees of freedom (not sample size), numeric of length 1.}

\item{x}{Population correlation coefficient, numeric of length 1 or more.
For some functions, squared correlation coefficients can be passed
as an argument when \code{do.square = FALSE}.}

\item{tol}{Tolerance used to judge wheter x is sufficiently close to 0 or +/- 1
(where numerical instability can happen).}

\item{tol.hg, maxiter}{Passed to \code{hgf()} as tol and maxiter there.}

\item{...}{Additional arguments in these functions are silently ignored.}

\item{do.square}{Whether \code{x} is to be squared before passed to \code{hgf()};
\code{TRUE} by default; set this \code{FALSE}
when \code{x} is already squared.}

\item{order.}{Order of asymptotic approximation (exponent \eqn{r} in \eqn{O(n^r)};
for asymptotic expressions only).}

\item{return_terms}{When TRUE, returns individual terms of different orders
(for asymptotic expressions only).
Retained mainly for debugging purposes, but could be used to
track the order in derived statistics.}

\item{R}{Correlation matrix; assumed to be validly constructed.}

\item{ind1, ind2}{Indices for the two focal coefficients (vectors of length 2).
Superseded by \code{i, j, k, l}.}

\item{i, j, k, l}{Alternative indices to specify the focal coefficients.
Superseded by \code{rij, rik, ril, rjk, rjl, rkl}.}

\item{rij, rik, ril, rjk, rjl, rkl}{The relevant population correlation coefficients.}
}
\description{
Internal functions to calculate exact and asymptotic moments of
sample correlation coefficients \eqn{r} from
population correlation coefficients.

\code{Exv.r1()}: exact expectation of \eqn{r}.

\code{Exv.r2()}: exact expectation of \eqn{r^2}.

\code{Exv.r3()}: exact expectation of \eqn{r^3}.

\code{Exv.r3d()}: alternative to \code{Exv.r3()}.

\code{Exv.r4()}: exact expectation of \eqn{r^4}.

\code{Exv.r4d()}: alternative to \code{Exv.r4()}.

\code{Var.r1()}: exact variance of \eqn{r}.

\code{Var.r2()}: exact variance of \eqn{r^2}.

\code{Cm3.r1()}: exact central third moment of \eqn{r}.

\code{Cm4.r1()}: exact central fourth moment of \eqn{r}.

\code{AExv.r1()}: asymptotic expectation of \eqn{r}.

\code{AVar.r1()}: asymptotic variance of \eqn{r}.

\code{ACm3.r1()}: asymptotic central third moment of \eqn{r}.

\code{ACm4.r1()}: asymptotic central fourth moment of \eqn{r}

\code{AExv.r2()}: asymptotic expectation of \eqn{r^2}

\code{AExv.r3()}: asymptotic expectation of \eqn{r^3}

\code{AExv.r4()}: asymptotic expectation of \eqn{r^4}

\code{AVar.r2()}: asymptotic variance of \eqn{r^2}

\code{ACov.r1()}: asymptotic covariance between two \eqn{r}'s.
}
\details{
All these functions except \code{ACov.r1()} requires the arguments
\code{n} and \code{x}. It can be \code{length(x) > 1}
(at least supeficially vectorized), but it is assumed \code{length(n) = 1}.

Covariance between two \eqn{r}'s is a function of
at least six population correlation coefficients.
\code{ACov.r1()} takes as an argument the entire population correlation
matrix \code{R} instead of individual correlation coefficients, and indices
for the focal correlation coefficients. Alternatively, the all six relevant
population correlation coefficients can directly be specified.

The exact expressions are from Soper et al. (1917) or Ghosh (1966).
In particular:
\itemize{
\item \code{Exv.r3()}: From Soper et al. (1917, eq. xxviii);
that of Ghosh (1966) seems inaccurate for this moment.
\item \code{Exv.r3d()}: Alternative to \code{Exv.r3()} from
Soper et al. (1917, eq. xxvi). Equivalent but slightly slower.
\item \code{Exv.r4()}: From Ghosh (1966, eq. 1).
\item \code{Exv.r4d()}: Alternative to \code{Exv.r4()} from
Soper et al. (1917, eq. xxi). Equivalent but slightly slower.
}

Some of the exact expressions might be instable when the \code{x} is small.

The asymptotic moments are from Ghosh (1966)
(note that the symbol \eqn{n} is used for sample size there).
\emph{Note}: The validity of \code{ACm3.r1()} may need critical evaluation,
as Ghosh's (1966) equation 1 (from which the result is supposedly drawn)
seems inaccurate (see Soper et al. [1917: eq. xxviii]).
The following functions depending on this might be inaccurate:
\code{AExv.r3()}, \code{AExv.r4()}, and \code{AVar.r2()}.

The asymptotic moment functions takes the argument \code{order.}.
This is the exponent in \eqn{O(n^r)}.
Higher orders do not always yield better approximations for ordinary
values of \eqn{n}. The allowed ranges and "good" choice
(empirically confirmed by comparison to exact expressions) are as follows:
\itemize{
\item \code{AExv.r1()},  \code{AExv.r2()}: 1--7. 2 is good
except when \eqn{n < 6} where 3 is better.
\item \code{AVar.r1()}: 1--7. 2 is good for \eqn{n} <100 or so;
for larger \eqn{n}, the difference tends to be negligible.
\item \code{ACm3.r1()}: 2--5. 3 is best for \eqn{n} > 15,
while 4 may work better for smaller \eqn{n}.
\item \code{ACm4.r1()}: 2--5. 3 is good for \eqn{n} > 20 especially when
\eqn{x} is small; 4 is better for small \eqn{n} and \eqn{x > 0.5}.
\item \code{AExv.r3()}: 1--5. 2 is good.
\item \code{AExv.r4()}: 1--5. 3 or 4 is good for small \eqn{x},
while 2 is better for \eqn{x > 0.4}.
\item \code{AVar.r2()}: 1--5. For \eqn{n < 100}, 3 or 4 is good for
very small \eqn{x}, while 2 is better for \eqn{x > 0.2}.
Slight overestimation seems to be common.
}

\code{ACov.r1()} is as in, e.g., Olkin & Siotani (1976), Olkin & Finn (1990).
When \code{ind1 == ind2}, it calculates the variance for the coefficient.

Multivariate normality is assumed for all moments.
Nevertheless, the distribution of \eqn{r} remains the same in many other
conditions (e.g., in all elliptically contoured distributions;
Anderson, 2003).
}
\examples{
# Trivial examples
n <- 10    # Which typically corresponds to sample size of 11
eigvaldisp:::Exv.r1(n, 0)
eigvaldisp:::Exv.r1(n, 0.5) # Underestimate
eigvaldisp:::Exv.r2(n, 0)   # This is n^-2 for x = 0

eigvaldisp:::AExv.r1(n, 0)
eigvaldisp:::AExv.r1(n, 0.5)
eigvaldisp:::AExv.r2(n, 0)  # Fairly close approximations of the above

# Comparison between exact vs. asymptotic expressions
n <- 3                 # Extremely small n
x <- seq(0, 1, 0.01)
cols <- rainbow(7)

# Exv.r1() vs AExv.r1() with different orders
plot(x, sapply(x, eigvaldisp:::Exv.r1, n = n), type = "l", lwd = 2,
     xlim = c(0, 1), ylim = c(0, 1))
for(i in 1:7) {
    lines(x, sapply(x, eigvaldisp:::AExv.r1, n = n, order = i),
          col = cols[i])
}
legend("topleft", title = "Order", legend = c(1:7, "Exact"),
       col = c(cols, "black"), lty = 1)

# Exv.r2() vs AExv.r2() with different orders
plot(x, sapply(x, eigvaldisp:::Exv.r2, n = n), type = "l", lwd = 2,
     xlim = c(0, 1), ylim = c(0, 1))
for(i in 1:7) {
    lines(x, sapply(x, eigvaldisp:::AExv.r2, n = n, order = i),
          col = cols[i])
}
legend("topleft", title = "Order", legend = c(1:7, "Exact"),
       col = c(cols, "black"), lty = 1)

}
\references{
Anderson, T. W. (2003). \emph{An Introduction to Multivariate Statistical
Analysis}, 3rd edition. John Wiley & Sons, Hoboken, New Jersey.

Ghosh, B. K. (1966). Expansions for the moments of the distribution
of correlation coefficients. \emph{Biometrika} \strong{53}, 258--262.
doi:\href{https://doi.org/10.2307/2334076}{10.2307/2334076}.

Olkin, I. & Finn, J. D. (1990). Testing correlated correlations.
\emph{Psychological Bulletin} \strong{108}, 330--333.
doi:\href{https://doi.org/10.1037/0033-2909.108.2.330}{10.1037/0033-2909.108.2.330}.

Olkin, I. & Siotani, M. (1976). Asymptotic distribution of functions of
a correlation matrix. Pp. 235--251 \emph{in} Editorial Committee for Publication
of Essays in Probability and Statistics, eds. \emph{Essays in Probability and
Statistics in Honor of Professor Junjiro Ogawa}. Shinko Tsusho, Tokyo.

Soper, H. E., Young, A. W., Cave, B. M., Lee, A. & Pearson, K. (1917).
On the distribution of the correlation coefficients in small samples.
Appendix II to the papers of ``Student'' and R. A. Fisher.
\emph{Biometrika} \strong{11}, 328--413.
doi:\href{https://doi.org/10.1093/biomet/11.4.328}{10.1093/biomet/11.4.328}.
}
\seealso{
\link{Exv.VXx}, \link{AVar.VRr_xx} for outer functions.
}
