% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exv.R
\name{Exv.VXax}
\alias{Exv.VXax}
\alias{Exv.VEav}
\alias{Exv.VRav}
\alias{Exv.VRar}
\alias{Var.VEav}
\alias{Var.VRav}
\alias{Var.VRar}
\title{Moments of ``bias-corrected'' eigenvalue dispersion indices}
\usage{
Exv.VEav(
  Sigma,
  n = 100,
  Lambda,
  drop_0 = FALSE,
  tol = .Machine$double.eps * 100,
  ...
)

Exv.VRav(
  Sigma,
  n = 100,
  Lambda,
  drop_0 = FALSE,
  tol = .Machine$double.eps * 100,
  ...
)

Exv.VRar(
  Rho,
  n = 100,
  Lambda,
  tol = .Machine$double.eps * 100,
  tol.hg = 0,
  maxiter.hg = 2000,
  ...
)

Var.VEav(
  Sigma,
  n = 100,
  Lambda,
  drop_0 = FALSE,
  tol = .Machine$double.eps * 100,
  ...
)

Var.VRav(
  Sigma,
  n = 100,
  Lambda,
  drop_0 = FALSE,
  tol = .Machine$double.eps * 100,
  ...
)

Var.VRar(
  Rho,
  n = 100,
  Lambda,
  tol = .Machine$double.eps * 100,
  tol.hg = 0,
  maxiter.hg = 2000,
  ...
)
}
\arguments{
\item{Sigma}{Population covariance matrix; assumed to be validly constructed.}

\item{n}{Degrees of freedom (not sample sizes); numeric of length 1 or more.}

\item{Lambda}{Numeric vector of population eigenvalues.}

\item{drop_0}{Logical, when \code{TRUE}, eigenvalues smaller than \code{tol} are dropped.}

\item{tol}{For covariance-related functions, this is the tolerance/threshold
to be used with drop_0. For correlation-related functions,
this is passed to \code{Exv.r2()} along with other arguments.}

\item{...}{In \code{Var.VRr()}, additional arguments are passed to an internal
function which it in turn calls. Otherwise ignored.}

\item{Rho}{Population correlation matrix; assumed to be validly constructed
(although simple checks are done).}

\item{tol.hg}{Passed to \code{Exv.r2()}; see description of that function.}

\item{maxiter.hg}{Passed to \code{Exv.r2()}; see description of that function.}
}
\value{
A numeric vector of the desired moment, corresponding to \code{n}.
}
\description{
Functions to calculate expectation/variance of eigenvalue dispersion indices
of covariance/correlation matrices.

\code{Exv.VEav()}: expectation of unbiased eigenvalue variance of
covariance matrix. Of little practical use because
this is just the population value \eqn{V(\Sigma)}.

\code{Exv.VRav()}: expectation of adjusted relative eigenvalue variance
of covariance matrix.

\code{Exv.VRar()}: expectation of adjusted relative eigenvalue variance
of correlation matrix.

\code{Var.VEav()}: variance of unbiased eigenvalue variance of
covariance matrix.

\code{Var.VR.av()}: variance of adjusted relative eigenvalue variance of
covariance matrix.

\code{Var.VRar()}: variance of adjusted relative eigenvalue variance
of correlation matrix.
}
\details{
Usage is identical to that of the corresponding unadjusted versions
(see \link{Exv.VXx}), which is in most cases called internally.
}
\examples{
# See also examples of Exv.VXx
# Covariance matrix
N <- 20
Lambda <- c(4, 2, 1, 1)
(Sigma <- GenCov(evalues = Lambda, evectors = "random"))
VE(S = Sigma)$VE
VE(S = Sigma)$VR
# Population values of V(Sigma) and Vrel(Sigma)

# Moments of bias-corrected eigenvalue variance of covariance matrix
Exv.VEav(Sigma, n = N - 1)
Var.VEav(Sigma, n = N - 1)
# The expectation is equal to the population value (as it should be)

# Moments of adjusted relative eigenvalue variance of covariance matrix
Exv.VRav(Sigma, n = N - 1)
Var.VRav(Sigma, n = N - 1)
# Slight underestimation is expected
# All these are the same with Lambda = Lambda is specified instead of Sigma.

# Correlation matrix
(Rho <- GenCov(evalues = Lambda / sum(Lambda) * 4, evectors = "Givens"))
VE(S = Rho)$VR
# Population value of Vrel(Rho), identical to Vrel(Sigma) as it should be

Exv.VRar(Rho, n = N - 1)
Var.VRar(Rho, n = N - 1)
# Slight underestimation is expected
# These results vary with the choice of eigenvalues
# If interested, repeat from the definition of Rho

# All options for Var.VRr() are accommodated
Var.VRar(Rho, n = N - 1, fun = "pfd") # Pan-Frank method; default
Var.VRar(Rho, n = N - 1, fun = "klv") # Konishi's theory

}
\references{
Watanabe, J. (2021). Statistics of eigenvalue dispersion indices:
quantifying the magnitude of phenotypic integration. \emph{Evolution},
doi:\href{https://doi.org/10.1111/evo.14382}{10.1111/evo.14382}.
}
\seealso{
\link{VXax} for ``bias-corrected'' estimators;
\link{Exv.VXx} for moments of unajusted versions.
}
