% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hgf}
\alias{hgf}
\title{Hypergeometric function wrapper}
\usage{
hgf(a1, a2, b1, x, tol = 0, maxiter = 2000)
}
\arguments{
\item{a1, a2}{Parameters in the numerator}

\item{b1}{Parameter in the denominator}

\item{x}{Argument of the hypergeometric function}

\item{tol, maxiter}{Passed to \code{genhypergeo} (and \code{hypergeo})}
}
\value{
A numeric vector corresponding to \code{x}
}
\description{
Wrapper for hypergeometric function with two upper and one lower arguments
(or the classic hypergeometric function).
}
\details{
This function utilizes the function \code{genhypergeo} from
the package \code{hypergeo}; when it appears to have failed,
then the function \code{hypergeo} is used.

Elements of \code{x} are assumed to be real numbers within \eqn{[-1, 1]}.
Complex arguments/returns are not assumed. No check is done.

Tries to reduce computational time by dropping duplicated values from
\code{x} while trying to retain its original \code{names} and \code{dim}.

This function primarily uses \code{hypergeo::genhypergeo()}, as this seems
to be more numerically stable than \code{hypergeo::hypergeo()} in some
conditions tested in development.
For the present purpose, \code{gsl::hyperg_2F1} may suffice,
but this has not been implemented.
}
\seealso{
\link[hypergeo]{genhypergeo}, \link[hypergeo]{hypergeo}
}
