% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_VE.R
\name{VXax}
\alias{VXax}
\alias{VEav}
\alias{VRav}
\alias{VRar}
\title{``Bias-corrected'' eigenvalue dispersion indices}
\usage{
VEav(
  X,
  S,
  L,
  n = N - as.numeric(center),
  divisor = c("UB", "ML"),
  m = switch(divisor, UB = N - 1, ML = N),
  center = TRUE,
  nv = 0,
  ...
)

VRav(
  X,
  S,
  L,
  n = N - as.numeric(center),
  divisor = c("UB", "ML"),
  m = switch(divisor, UB = N - 1, ML = N),
  center = TRUE,
  nv = 0,
  ...
)

VRar(X, S, L, n = N - as.numeric(center), center = TRUE, nv = 0, ...)
}
\arguments{
\item{X}{Data matrix from which covariance/corrrelation matrix is obtained.}

\item{S}{Covariance/correlation matrix.}

\item{L}{Vector of eigenvalues.}

\item{n}{Degrees of freedom; required unless X is provided.}

\item{divisor}{Either \code{"UB"} (default) or \code{"ML"},
to decide the default value of \code{m}.}

\item{m}{Divisor for the sample covariance matrix (\eqn{n*} in Watanabe (2021)).}

\item{center}{Logical to specify whether sample-mean-centering should be done.}

\item{nv}{Numeric. Specify how many eigenvectors are to be retained; default 0.}

\item{...}{Arguments \code{sub}, \code{drop_0}, \code{tol} can be
passed to \code{VE()}. Other arguments will not influence the result.}
}
\value{
A list similar to the output of \code{VE()}, with an additional element:
\describe{
\item{\code{VEav()}}{\code{$VEav}: Bias-corrected eigenvalue
variance of covariance matrix.}
\item{\code{VRav()}}{\code{$VRav}: Adjusted relative
eigenvalue variance of covariance matrix}
\item{\code{VRar()}}{\code{$VRar}: Adjusted relative
eigenvalue variance of correlation matrix}
}
}
\description{
These functions calculate bias-corrected or adjusted
eigenvalue dispersion indices

\code{VEav}: bias-corrected eigenvalue variance of covariance matrix.

\code{VRav}: adjusted relative eigenvalue variance of covariance matrix.

\code{VRar}: adjusted relative eigenvalue variance of correlation matrix.
}
\details{
These functions are to be used with sample data matrix,
covariance/correlation matrix, or eigenvalues, and not to be used with
population quantities. Unless \code{X} is provided, the degrees of freedom
\code{n} should be specified as this is required for adjustment.

These functions internally call \code{VE(L = L, nv = 0, ...)} with
appropriately constructed (or user-specified) \code{L},
and adjusted eigenvalue dispersion indices are appended to the outcome list.
If \code{nv > 0}, eigenvectors are calculated before this function call
and returned as well.

Bias correction is possible for eigenvalue variance of covariance matrices,
but not straightforward for relative eigenvalue variance (of either
covariance or correlation matrices), because the latter is a nonlinear
function of the population value (see Watanabe, 2021).
Although this could potentially be achievable for correlation matrices,
expression for its variance is not known to the author.

The bias-corrected eigenvalue variance of a sample covariance matrix
also has smaller variance than the uncorrected version.
On the other hand, the adjusted relative eigenvalue variance of a
sample covariance/correlation matrix has larger variance than
the unadjusted version.
}
\examples{
# Spherical covariance matrix (VE = VR = 0)
Sigma <- diag(4)
VE(S = Sigma)

N <- 20
set.seed(375)
X <- eigvaldisp:::rmvn(N = N, Sigma = Sigma)
S <- cov(X)
L <- eigen(S)$values

# Bias-corrected eigenvalue variance of covariance matrix
VEav(X = X)
VEav(S = S, n = N - 1)$VEav # Same, but n is required
VEav(L = L, n = N - 1)$VEav # Same
# Note the overestimation in VE and VR,
# and (slightly) better performance of VEav
# (although it does not always work this well)

# Adjusted relative eigenvalue variance of covariance matrix
VRav(X = X)$VRav

# Population value for the correlation matrix (same to Sigma in this case)
VE(S = stats::cov2cor(Sigma))

# Adjusted relative eigenvalue variance of correlation matrix
VRar(X = X)


# Covariance/correlation matrix with strong covariation (VR = 0.8)
(Sigma2 <- GenCov(p = 4, VR = 0.8, evectors = "Givens"))
VE(S = Sigma2)

N <- 20
X2 <- eigvaldisp:::rmvn(N = N, Sigma = Sigma2)

VEav(X = X2)
# This is better than the un-corrected version

VRav(X = X2)
VRar(X = X2)
# But these are slightly worse (as expected)

}
\references{
Watanabe, J. (2021). Statistics of eigenvalue dispersion indices:
quantifying the magnitude of phenotypic integration. \emph{Evolution},
doi:\href{https://doi.org/10.1111/evo.14382}{10.1111/evo.14382}.
}
\seealso{
\link{VE} for the main function;
\link{Exv.VXx} for expectation (bias) in the ordinary estimators;
\link{Exv.VXax} for the expectation/variance of the adjusted estimators.
}
